add ui updates
