+++ src/Components/BookDetails/BookDetails.css
@@ -34,7 +34,7 @@
   padding: 20px 10px;
   background-position: center;
   width: auto;
-  height: 60vh;
+  height: 70vh;
   background-color: #E8E8E8;
   position: relative;
 }
@@ -86,8 +86,21 @@
   transition: background-color 0.3s ease; 
   font-family: 'Gill Sans', 'Gill Sans MT', 'Calibri', 'Trebuchet MS', sans-serif; /* Choose a font for the button */
 }
-
+.addButton1 {
+  background-color: #45C793;
+  color: white;
+  border: none; 
+  padding: 10px 20px; 
+  border-radius: 25px;
+  font-size: 16px; 
+  cursor: pointer; 
+  transition: background-color 0.3s ease; 
+  font-family: 'Gill Sans', 'Gill Sans MT', 'Calibri', 'Trebuchet MS', sans-serif; /* Choose a font for the button */
+}
 .addButton:hover {
+  background-color: #28a745; 
+}
+.addButton1:hover{
   background-color: #28a745; 
 }
 .removeButton {
@@ -104,4 +117,36 @@
 }
 .removeButton:hover {
   background-color: #911731;
+}
+.modal-content{
+  display: flex;
+  justify-content: left;
+  align-items: center;
+  gap: 10px;
+}
+.modalOpt{
+  padding: 1px 10px;
+  height: 40px;
+  border: 1px solid #ccc;
+  border-radius: 100px;
+  background-color: #fff;
+  font-size: 16px;
+  cursor: pointer;
+  appearance: none; /* Removes default arrow for custom styling */
+  outline: none;
+  transition: border-color 0.3s ease;
+}
+.closeButton{
+  background-color: white;
+  color: black;
+  border: 1px; 
+  padding: 10px 20px; 
+  border-radius: 25px;
+  font-size: 16px; 
+  cursor: pointer; 
+  transition: background-color 0.3s ease; 
+  font-family: 'Gill Sans', 'Gill Sans MT', 'Calibri', 'Trebuchet MS', sans-serif;
+}
+.closeButton:hover{
+  background-color: #ccc;
 }
