+++ src/Components/Header/Header.js
@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
+import { useHandleLogout } from '../utils/handleUtil';
 import './Header.css';
 
-const Header = ({ user, handleLogout, showWelcome = true, showSignUp = true, showSearch = true, showMyLists = true, showProfile = true }) => {
+const Header = ({ user, showWelcome = true, showSignUp = true, showSearch = true, showMyLists = true, showProfile = true }) => {
   const navigate = useNavigate();
   const [hasWishlists, setHasWishlists] = useState(false);
-
+  const handleLogout = useHandleLogout();
   useEffect(() => {
     if (user) {
       const fetchWishlists = async () => {
