+++ src/Components/ListPage/ListPage.css
@@ -11,6 +11,9 @@
     display: flex;
     gap: 10px;
 }
+.wishlist-section{
+  background-color: white;
+}
 .Logout{
     display: flex;
     justify-content: center;
@@ -81,6 +84,20 @@
     object-fit: contain;
     margin-top: 60px;
     image-rendering: crisp-edges;
+  }
+  .createButton{
+    background-color: white;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    overflow: hidden;
+    text-align: center;
+    transition: transform 0.3s ease-in-out;
+    padding: 15px;
+    height: 100%;
+  }
+  .createButton:hover {
+    transform: scale(1.05);
   }
   .main-layout{
     background-color: #F0F0F0;
