+++ src/Components/SearchPage/SearchPage.css
@@ -79,7 +79,11 @@
     font-weight: 600;
     font-size: 14px;
     cursor: pointer;
+    transition: background-color 0.3s ease;
     border: none;
+}
+.SearchSubmit:hover{
+  background-color: #3ca882;
 }
 .about-description {
     text-align: left;
